private_endpoint_regional_mode
