+++ internal/service/privateendpointregionalmode/data_source_private_endpoint_regional_mode.go
@@ -1,4 +1,4 @@
-package mongodbatlas
+package privateendpointregionalmode
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
 )
 
-func DataSourcePrivateEndpointRegionalMode() *schema.Resource {
+func DataSource() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceMongoDBAtlasPrivateEndpointRegionalModeRead,
 		Schema: map[string]*schema.Schema{
