+++ internal/service/privateendpointregionalmode/data_source_private_endpoint_regional_mode_test.go
@@ -1,4 +1,4 @@
-package mongodbatlas_test
+package privateendpointregionalmode_test
 
 import (
 	"context"
